Moved getData to its own componen
